removing some unnecessary changes
